resource holon variavel sequence creation fixed
initially four variables were created in a single transiction, but the creation order can be different from one simulation to other and the relation between the variable name in the transition and the arc is not maintained, therefor the execution result on java exceptions. Since, in a transition, is possible to create multiple pointers to the same instance, this has been used thus allow to increase complexity in the  model, to fix the issue, but maintaining the same number of instants and therefor not running into memory issues.